+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/DataAccess/BugRepository.java
@@ -1,6 +1,6 @@
 package com.elsys.globalserver.DataAccess;
 
-import com.elsys.globalserver.DatabaseEntities.Bug;
+import com.elsys.globalserver.Models.Bug;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BugRepository extends CrudRepository<Bug, Integer> {
