+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/DataAccess/MedicineRepository.java
@@ -1,6 +1,6 @@
 package com.elsys.globalserver.DataAccess;
 
-import com.elsys.globalserver.DatabaseEntities.Medicine;
+import com.elsys.globalserver.Models.Medicine;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MedicineRepository extends CrudRepository<Medicine, Integer> {
