+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/DataAccess/UserRepository.java
@@ -1,6 +1,6 @@
 package com.elsys.globalserver.DataAccess;
 
-import com.elsys.globalserver.DatabaseEntities.User;
+import com.elsys.globalserver.Models.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
