+++ GlobalServerSoftware/src/main/java/com/elsys/globalserver/Services/PrescriptionsService.java
@@ -3,9 +3,9 @@ package com.elsys.globalserver.Services;
 import com.elsys.globalserver.DataAccess.MedicineRepository;
 import com.elsys.globalserver.DataAccess.PrescriptionRepository;
 import com.elsys.globalserver.DataAccess.UserRepository;
-import com.elsys.globalserver.DatabaseEntities.Medicine;
-import com.elsys.globalserver.DatabaseEntities.Prescription;
-import com.elsys.globalserver.DatabaseEntities.User;
+import com.elsys.globalserver.Models.Medicine;
+import com.elsys.globalserver.Models.Prescription;
+import com.elsys.globalserver.Models.User;
 import com.elsys.globalserver.Exceptions.Medicines.MedicineNotFoundException;
 import com.elsys.globalserver.Exceptions.Prescriptions.PrescriptionNotFoundException;
 import com.elsys.globalserver.Exceptions.Users.DoctorNotFoundException;
